Move php extractors to a TokenExtractor sub namespace
